Refactor import paths for prisma and user utilities
