+++ src/actions/acceptFriendRequest.ts
@@ -1,7 +1,7 @@
 "use server";
 
-import prisma from "@/lib/prisma";
-import { validateUser } from "@/util/userUtil";
+import prisma from "@/util/prisma";
+import { validateUser } from "@/util/user";
 
 export default async function acceptFriendRequestAction(id: string) {
     try {
