+++ src/actions/sendFriendRequest.ts
@@ -1,7 +1,7 @@
 "use server";
 
-import prisma from '@/lib/prisma';
-import { validateUser } from '@/util/userUtil';
+import prisma from '@/util/prisma';
+import { validateUser } from '@/util/user';
 
 export async function sendFriendRequest(friendId: string) {
     try {
