+++ src/app/layout.tsx
@@ -8,6 +8,7 @@ import { Roboto } from "next/font/google";
 const roboto = Roboto({
   weight: "500",
   display: "swap",
+  subsets: ["latin"],
 });
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
